requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
ollama==0.0.11
textblob==0.17.1
langdetect==1.0.9
pandas==2.1.1
numpy==1.24.3
